# back/api/swagger.yaml
openapi: "3.0.1"
info:
  title: PetProject
  description: PetProject
  version: 1.0.0
paths:
  /users/me:
    get:
      summary: "Information about current user"
      description: "Get information about current user that made request"
      operationId: "userMe"
      security:
        - JWTAuth: []

      responses:
        200:
          description: "Success response"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
                  error:
                    $ref: '#/components/schemas/Error'
  /users/register:
    post:
      summary: "Register user"
      description: "Register user with given values"
      operationId: "registerUser"
      responses:
        200:
          description: "Success response"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
                  error:
                    $ref: '#/components/schemas/Error'
      requestBody:
        $ref: '#/components/requestBodies/RegisterUser'
  /users/login:
    post:
      summary: "Login user"
      description: "Make user login"
      operationId: "loginUser"
      responses:
        200:
          description: "Success response"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Authorization'
                  error:
                    $ref: '#/components/schemas/Error'
      requestBody:
        $ref: '#/components/requestBodies/LoginUser'

components:
  schemas:
    User:
      description: "User schema"
      type: object
      properties:
        username:
           $ref: '#/components/schemas/Username'
        email:
          $ref: '#/components/schemas/Email'
      required:
        - username
        - email
    Username:
      description: "Username of user"
      type: string
      x-oapi-codegen-extra-tags:
        validate: "required"
    Password:
      description: "Password of user"
      type: string
      x-oapi-codegen-extra-tags:
        validate: "required"
    Email:
      description: "Email of user"
      type: string
      x-oapi-codegen-extra-tags:
        validate: "required,email"
    Error:
      description: "Error that appeared while processing request"
      type: string
    AccessToken:
      description: "Access token"
      type: string
    Authorization:
      description: "Authorization info"
      type: object
      properties:
        access_token:
          $ref: '#/components/schemas/AccessToken'
      required:
        - access_token
  securitySchemes:
    JWTAuth:
      type: apiKey
      in: header
      scheme: bearer
      name: Authorization
  requestBodies:
    RegisterUser:
      description: "Body to register user"
      content:
        application/json:
          schema:
            type: object
            properties:
              username:
                $ref: '#/components/schemas/Username'
              email:
                $ref: '#/components/schemas/Email'
              password:
                $ref: '#/components/schemas/Password'
              repeat_password:
                description: "Repeat password"
                type: "string"
                x-oapi-codegen-extra-tags:
                  validate: "required,eqfield=Password"
            required:
              - username
              - email
              - password
              - repeat_password
    LoginUser:
      description: "Login user"
      content:
        application/json:
          schema:
            type: object
            properties:
              username:
                $ref: '#/components/schemas/Username'
              password:
                $ref: '#/components/schemas/Password'
            required:
              - username
              - password